+++ library/inspector/src/main/java/io/mehow/laboratory/inspector/FeatureCoordinates.kt
@@ -1,0 +1,6 @@
+package io.mehow.laboratory.inspector
+
+internal data class FeatureCoordinates(
+  val sectionIndex: Int,
+  val featureIndex: Int,
+)