+++ library/inspector/src/main/java/io/mehow/laboratory/inspector/SectionFragment.kt
@@ -1,0 +1,57 @@
+package io.mehow.laboratory.inspector
+
+import android.os.Bundle
+import android.view.View
+import androidx.core.os.bundleOf
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.activityViewModels
+import androidx.lifecycle.lifecycleScope
+import androidx.recyclerview.widget.RecyclerView
+import io.mehow.laboratory.Feature
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.launch
+
+internal class SectionFragment : Fragment(R.layout.io_mehow_laboratory_feature_group) {
+  val inspectorViewModel by activityViewModels<InspectorViewModel> {
+    InspectorViewModel.Factory(LaboratoryActivity.configuration, searchViewModel)
+  }
+  private val searchViewModel by activityViewModels<SearchViewModel> { SearchViewModel.Factory }
+  private val sectionName get() = requireStringArgument(sectionKey)
+
+  private lateinit var layoutManager: SmoothScrollingLinearLayoutManager
+  private val featureAdapter = FeatureAdapter(object : FeatureAdapter.Listener {
+    override fun onSelectFeature(feature: Feature<*>) = inspectorViewModel.selectFeature(feature)
+
+    override fun onGoToFeature(feature: Class<Feature<*>>) {
+      lifecycleScope.launch { inspectorViewModel.goTo(feature) }
+    }
+  })
+
+  override fun onViewCreated(view: View, inState: Bundle?) {
+    view.findViewById<RecyclerView>(R.id.io_mehow_laboratory_feature_section).apply {
+      layoutManager = SmoothScrollingLinearLayoutManager(requireActivity()).also {
+        this@SectionFragment.layoutManager = it
+      }
+      adapter = featureAdapter
+      hideKeyboardOnScroll()
+    }
+    observeGroup()
+  }
+
+  private fun observeGroup() = inspectorViewModel.sectionFlow(sectionName)
+      .onEach { featureAdapter.submitList(it) }
+      .launchIn(viewLifecycleOwner.lifecycleScope)
+
+  fun scrollTo(index: Int) = layoutManager.smoothScrollTo(index)
+
+  companion object {
+    private const val sectionKey = "Section.Key"
+
+    fun create(section: String): SectionFragment {
+      return SectionFragment().apply {
+        arguments = bundleOf(sectionKey to section)
+      }
+    }
+  }
+}